refactor(tui): extract run flow and agent event handling component
